// src/apiV2/a7-service/model/postPhotosAddlayerBodyOutputResolution.ts
/**
 * Generated by orval v6.6.2 🍺
 * Do not edit manually.
 * WanmaX Backend API
 * API документация для WanmaX Backend - Автоматизированный конвейер обработки фотографий
 * OpenAPI spec version: 1.0.0
 */

/**
 * Разрешение выходного изображения
 */
export type PostPhotosAddlayerBodyOutputResolution = '2K' | '4K';


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PostPhotosAddlayerBodyOutputResolution = {
  '2K': '2K' as PostPhotosAddlayerBodyOutputResolution,
  '4K': '4K' as PostPhotosAddlayerBodyOutputResolution,
};