// src/apiV2/a7-service/model/photoStatus.ts
/**
 * Generated by orval v6.6.2 🍺
 * Do not edit manually.
 * WanmaX Backend API
 * API документация для WanmaX Backend - Автоматизированный конвейер обработки фотографий
 * OpenAPI spec version: 1.0.0
 */

/**
 * Статус обработки фото (improve/addlayer) для polling-индикации
 */
export type PhotoStatus = 'idle' | 'processing' | 'success' | 'failed';


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PhotoStatus = {
  idle: 'idle' as PhotoStatus,
  processing: 'processing' as PhotoStatus,
  success: 'success' as PhotoStatus,
  failed: 'failed' as PhotoStatus,
};