// src/apiV2/a7-service/model/photoStatusOperation.ts
/**
 * Generated by orval v6.6.2 🍺
 * Do not edit manually.
 * WanmaX Backend API
 * API документация для WanmaX Backend - Автоматизированный конвейер обработки фотографий
 * OpenAPI spec version: 1.0.0
 */

/**
 * Какая операция запущена / упала последней
 */
export type PhotoStatusOperation = 'improve' | 'addlayer' | null;


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PhotoStatusOperation = {
  improve: 'improve' as PhotoStatusOperation,
  addlayer: 'addlayer' as PhotoStatusOperation,
};